Rudimentary geom checking
